requirements.txt
Django==1.7.1
django-appconf==0.6
django-bootstrap3==4.11.0
django-bower==5.0.1
django-compressor==1.4
django-debug-toolbar==1.2.2
django-filter==0.8
django-grappelli==2.6.2
django-taggit==0.12.2
django-test-pep8==0.1
djangorestframework==2.4.4
mysqlclient==1.3.4
pep8==1.5.7
six==1.8.0
sqlparse==0.1.13
